refactor(widgets): restructure folder layout and remove dead code
Restructure widgets folder for cleaner organization:
- Move ListWidget and LabelWidget from widgets/widgets/ to widgets/
- Move Renderable interface from base/ to widgets/ root
- Move TooltipRenderer from util/ to core/
- Delete unused CachedComponent class

Update all imports in dependent modules to reflect new paths.
