+++ src/main/java/com/github/lutzluca/btrbz/core/order_book/OrderBookScreen.java
@@ -6,8 +6,8 @@ import java.util.ArrayList;
 import java.util.List;
 import com.github.lutzluca.btrbz.data.OrderModels.OrderType;
 import com.github.lutzluca.btrbz.widgets.core.WidgetManager;
-import com.github.lutzluca.btrbz.widgets.widgets.ListWidget;
-import com.github.lutzluca.btrbz.widgets.base.Renderable;
+import com.github.lutzluca.btrbz.widgets.ListWidget;
+import com.github.lutzluca.btrbz.widgets.Renderable;
 import net.hypixel.api.reply.skyblock.SkyBlockBazaarReply.Product.Summary;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
